Extract `#on_heredoc` callback into a mixin
This is an intermediate extraction. Some of the functionality will be
moved into a `StringNode` extension, and the callback will be moved
after that.
