+++ lib/rubocop.rb
@@ -111,6 +111,7 @@ require 'rubocop/cop/mixin/min_body_length'
 require 'rubocop/cop/mixin/multiline_expression_indentation'
 require 'rubocop/cop/mixin/multiline_literal_brace_layout'
 require 'rubocop/cop/mixin/negative_conditional'
+require 'rubocop/cop/mixin/heredoc'
 require 'rubocop/cop/mixin/on_normal_if_unless'
 require 'rubocop/cop/mixin/parentheses'
 require 'rubocop/cop/mixin/parser_diagnostic'
